fix: :bug: fix bug in vec_sort, optimise my_swap + my_fmt
vec_sort: heap-overlow due to size_t -1 ....

+++ lib/my_vec/sort.c
@@ -13,15 +13,14 @@ static int default_compare(void const *a, void const *b)
     return *(int *)a - *(int *)b;
 }
 
-static size_t partition(
-    vec_t *v, int (*cmp_fn)(void const *, void const *), size_t start,
-    size_t end
+static long partition(
+    vec_t *v, int (*cmp_fn)(void const *, void const *), long start, long end
 )
 {
-    size_t pivot = end;
-    size_t i = start - 1;
+    long pivot = end;
+    long i = start - 1;
 
-    for (size_t j = start; j < end; ++j) {
+    for (long j = start; j < end; ++j) {
         if (cmp_fn(vec_at(v, j), vec_at(v, pivot)) <= 0) {
             i++;
             my_swap(vec_at(v, i), vec_at(v, j), v->_elem_size);
@@ -33,14 +32,13 @@ static size_t partition(
 }
 
 static void quick_sort(
-    vec_t *v, int (*cmp_fn)(void const *, void const *), size_t start,
-    size_t end
+    vec_t *v, int (*cmp_fn)(void const *, void const *), long start, long end
 )
 {
     if (start >= end)
         return;
 
-    size_t center = partition(v, cmp_fn, start, end);
+    long center = partition(v, cmp_fn, start, end);
 
     quick_sort(v, cmp_fn, start, center - 1);
     quick_sort(v, cmp_fn, center + 1, end);
